AppContainer: provide both a functor-based and a functor-less version of the binding.
