+++ src/AppContainer.bs.js
@@ -1,11 +1,1 @@
-'use strict';
-
-var ReactNavigation = require("react-navigation");
-
-function Make(S) {
-  var make = ReactNavigation.createAppContainer(S[/* navigator */0]);
-  return /* module */[/* make */make];
-}
-
-exports.Make = Make;
-/* react-navigation Not a pure module */
+/* This output is empty. Its source's type definitions, externals and/or unused code got optimized away. */
