+++ src/AppContainer.re
@@ -1,3 +1,17 @@
+// Low-level, zero-cost bindings to AppContainer.
+//
+// Usage:
+//
+// let appContainer = AppContainer.makeAppContainer(rootNavigator);
+//
+// React.createElement(
+//   appContainer,
+//   AppContainer.makeProps(
+//     ~screenProps, /* ... more props ... */
+//     (),
+//   ),
+// );
+
 type persistNavigationState = NavigationState.t => Js.Promise.t(unit);
 type loadNavigationState = unit => Js.Promise.t(option(NavigationState.t));
 
@@ -9,27 +23,20 @@ type appContainerProps('screenProps) = {
   "setNavigatorRef": Js.Nullable.t(NavigationContainer.t) => unit,
 };
 
-module Make = (S: {
-                 type screenProps;
-                 let navigator: Navigator.t;
-               }) => {
-  [@bs.module "react-navigation"]
-  external make:
-    Navigator.t => React.component(appContainerProps(S.screenProps)) =
-    "createAppContainer";
+[@bs.obj]
+external makeProps:
+  (
+    ~persistNavigationState: persistNavigationState=?,
+    ~loadNavigationState: loadNavigationState=?,
+    ~screenProps: 'screenProps=?,
+    ~setNavigatorRef: Js.Nullable.t(NavigationContainer.t) => unit=?,
+    ~key: string=?,
+    unit
+  ) =>
+  appContainerProps('screenProps) =
+  "";
 
-  [@bs.obj]
-  external makeProps:
-    (
-      ~persistNavigationState: persistNavigationState=?,
-      ~loadNavigationState: loadNavigationState=?,
-      ~screenProps: S.screenProps=?,
-      ~setNavigatorRef: Js.Nullable.t(NavigationContainer.t) => unit=?,
-      ~key: string=?,
-      unit
-    ) =>
-    appContainerProps(S.screenProps) =
-    "";
-
-  let make = make(S.navigator);
-};
+[@bs.module "react-navigation"]
+external makeAppContainer:
+  Navigator.t => React.component(appContainerProps('screenProps)) =
+  "createAppContainer";
