+++ src/AppContainerFunctor.bs.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var ReactNavigation = require("react-navigation");
+
+function Make(S) {
+  var make = ReactNavigation.createAppContainer(S[/* navigator */0]);
+  return /* module */[/* make */make];
+}
+
+exports.Make = Make;
+/* react-navigation Not a pure module */
